readme and refactor
