+++ aeon_neuro/distances/__init__.py
@@ -2,7 +2,7 @@
 
 __all__ = ["affine_invariant_distance", "log_euclidean_distance"]
 
-from aeon_neuro.distances._reimannian import (
+from aeon_neuro.distances._riemannian import (
     affine_invariant_distance,
     log_euclidean_distance,
 )
