+++ aeon_neuro/distances/_riemannian.py
