+++ aeon_neuro/distances/tests/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["affine_invariant_distance"]
 
-from aeon_neuro.distances._reimannian import affine_invariant_distance
+from aeon_neuro.distances._riemannian import affine_invariant_distance
